Integration test added.
