+++ test/infrastructure/GtMotive.Estimate.Microservice.InfrastructureTests/GlobalSuppressions.cs
@@ -1,0 +1,8 @@
+﻿// This file is used by Code Analysis to maintain SuppressMessage
+// attributes that are applied to this project.
+// Project-level suppressions either have no target or are given
+// a specific target and scoped to a namespace, type, member, etc.
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Naming", "CA1707:Identifiers should not contain underscores", Justification = "Tests name could use underscores", Scope = "member", Target = "~M:GtMotive.Estimate.Microservice.InfrastructureTests.Vehicles.CreateVehicleTests.Create_ShouldAddNewVehicleToDatabase_WhenRequestIsCorrect~System.Threading.Tasks.Task")]
