+++ test/infrastructure/GtMotive.Estimate.Microservice.InfrastructureTests/appsettings.Testing.json
@@ -1,0 +1,6 @@
+﻿{
+  "AllowedHosts": "*",
+  "AppSettings": {
+    "JwtAuthority": "https://identity.mygtmotive.com"
+  }
+}
